Add driver for EtherCAT based motion controllers
Add a demo driver for EtherCAT based system, curently tested against the
simulator.
In the longer run, it should replace the driver from MCAG_Base_Project.

Change the hard- and soft limits in the simulator:
We currently home against the low limit switch, which means
that the position of the simulated limit switches needs to be updated.