test 090: Work around missing sync of .RBV and .VAL at startup
When the record is initialized in init_record(), the .VAL field
should be synchronized with the .RBV field.

This does not always work for asyn motors, since the .RBV may have not
been retrieved yet (it will come in a callback later).

The consequence is that .VAl is 0 and stays there, while .RBV is e.g. 100.1
This can be observed in the simulator.

When writing to the .TWV/.TWR field, the value of .VAL is taken to calculate
the position after the tweak, which is 1.1 instead of 102.1
The right place to fixing this is the record. This needs more work and testing
and will be done in a later commit.

As a workaround, change test 090 to read the .RBV field and use .VAL field
for the tweaking. (Remember that test 090 is there to test the commissioning
of the hardware, encoder and motor scalings and polarity)