EthercatMC: Treat a motor with power off as done.
Sometimes the "bBusy" bit goes not to 0 and DMOV
will never become true.

The system is locked up.
Make a possible workaround:
The use can switch off the amplifier and press the HOMF button.
The home sequence will start and when finished, the DMOV will become true.

(This is a workaround, the bBusy issure should be fixed in the MCU
 software)