EthercatMC: Acc-RB and Dec-RB
We read back the configure acceleration and deceleration from
the controller.
These are not the actual (and measured) values,
but "read back values" in the sense "this is what we have send to
the controller and now we read back the set point".
Fix this according to the ESS naming convention.