EthercatMCAxis: Stop when switching of the amplifier
In positioning mode, the bExecute=1 is saying "hold position".
When the amplifier is off and the motor is e.g. moved manually,
then the movement may be out of "in target position range".
As a result, bBusy goes high.
Avoid this, and stop execution prior to switching of the amplfier.