simulator/EtherCAT: Simulate amplifier on failure
Some old code already allowed to simulate the amplifier can
not be switched on. Improve it to be part of hw_motor:
getAmplifierLockedToBeOff() and setAmplifierLockedToBeOff() allow
to amplifer to be locked, which will turn it off and keep it off.

Move the commands into cmd_Sim.c to allow
Sim.M1.bAmplifierLockedToBeOff=0.