SimAxis: Different ways to simulate an encoder
Document the different ways to simulate an encoder, especially
using an EtherCAT terminal with an incremental encoder.
This will give us the 16 bit raw value, including wrap-arounds.
So reading the wrapping value does not give a kind of absolute
position, but allows to look at the mechanics (does it move 5
encoder ticks or not)