EthercatMC: Move adsport into drvlocal
Move the adsport into drvlocal.
The adsport defines which ADS device we communicate with, and so
far we always had 501 in our project.
So we had "501" spread out all over the code.
Collect the port at one place, to be able to change it later.

Note/TODO:
We still have "ADSPORT=852/.THIS.sFeatures?" in the code.