Read more parameters from controller (part1)
Some controllers are fully configured, when they are connected
to the EPICS IOC (and the axisRecord is loaded).
These controllers have already settings for e.g. velocity,
acceleration, maximum velocity and others.

At the same time the reocord has fields for VELO, JVEL, JAR, VMAX and
others, which now may be conflict with the controller values.
Or, if the conflict should be avoided, they must be kept in sync.

In the future, a workflow like this should be possible:
- The motion controller is fully configured and tested using the
  engineering tool supplied by the vendor.
- The dynamic parameters are all trimmed, soft limits defined.
- The EPICS IOC is connected, and values are read from the controller
  and copied into the record.

This is the first commit:
- Define parameters for moving velocity, maximum velocity,
  jogging velocity, acceleration and forward them into device support.
  Currently the Ethercat based driver uses those.
- Define parameters for "set point deadband" and "retry deadband",
  they are currently not used.