Revert "EthercatMC: Remove mres handling"
This reverts commit b422e0747aeee723139f21016bbd62ec874b08b1.
Re-introduce the MRES compensation in the driver, as old Db files
rely on it.
One thing to be done in the next commit is to scale the values
for e.g. the "read only" soft limts.
The next thing to do is to allow an overwrite from st.cmd, as test case
900 now fails.