EthercatMCaxis*.opi: Change the back ground color
The background color was some kind of grey, which made it hard to read the
values. Change the background color from grey to white.

Because css would had changed the font and would had re-written the whole file,
the change was done with a text editor.

The real diff looks like this:
 -    <background_color>
 -      <color name="IO OutputPV Bg" red="121" green="123" blue="126" />
 -    </background_color>

 +    <background_color>
 +      <color red="255" green="255" blue="255" />
 +    </background_color>