Don't use dirty bits; Improve axis not homed
Remove the usage of the dirty bits:
Methods now return a status and initialPoll() may fail instead.
(In this case will the poller call initialPoll() again)

Fix a bug when "axis not homed" is a problem or not:
Only if there are softlimits and if there is a homing procedure defined,
"not homed" is reported as a problem.