The raw Encode value is not useful for REP
When there is an encoder, the user may want to see the value on the display.
So far the raw value had been put into the REP field, but this is not ideal
when an incremental encoder is used.
When the system is powered up, 0 in the encoder may be different from 0
at the axis, once the axis is homed.
The axis record assumes that both have the same 0.

The motor controller delivers the correct position in fActValue,
so put fActValue into motorEncoderPosition and define a new ai record:
"-EncAct", which is shown in CSS.

Remove all ERES in the start scripts.