Simulator: Improve power on and AmplifierLockedToBeOff
hw_motor: Improve the ramp-up of amplifier

When the power was turned on, it may take some time until
the amplifier is ready. Typically 0.5 .. 3 seconds.
When the amplifier is turned on from cmd_EAT, it is set to 95%
and the ramped up with 1% units / second.
This ramp up was depending in the polling, each polling
assumed to take 1 second.
This simple logic does not work, when the motor is not polled,
since the driver does a blind "sleep" for 5 seconds.

Solution:
Use the elapsed time to raise the power to 100%.

Make sure that AmplifierLockedToBeOff is used, this is part of test 210.