Simulator: Add getMotorPosStepped()
Factor out the re-callculation of the actual position:
It may be the same as the simulator has, or it may
be re-calculated to a full step.
The full step calculation is skipped, when the scaling
is 0.0, to be more exact when "urev" is 0.0

The exact position is needed in TC 142.