Add Homing Procedure read back PV and OPI
Make it possible to look at the configured homing procedure:
Introduce a new PV, $(P)$(M)-ProcHom-RB which is read only
and in text form. (Use a mbbbi record).
Update the Config opi to display the PV.

Note: On the long term, the whole EthercatMChome.template needs to
go away, and this is the first step.