Remove EthercatMChome.template
Remove all the parameters from the driver and the whole home.template.
From now on, the homing parameters should be retrieved from the controller.
This is not done yet, today the ProcHom= parameter can be specified
when loading the driver.
On the longer run, the homing procedure should be read per axis from the
controller.