Ad startup files for SlitMCU010
Add start scripts which allow to commission a dual axis crate
with incremental encoders in a "slit way":
Position 0 is in the middle of the axis, in the sense that the
edge of the wagon(s) simulates the edge of the blades.

Moving both axes to postion 0 will look like a closed slit at postion 0,
which is in the middle of the beam.