Split EthercatMCApp/ into EthercatMCApp/ and Sup
Two things (and more) are generated by the Makefile:
A library and an executable library.
In a typical EPICS application there are 2 directories with Makefile,
like "xxxApp" and "xxExApp" or (better) "xxxSup" and "xxxTestApp"

In order to build and run an application which uses the calc module,
move out source code and templates into EthercatMcSup and keep the source
for the "test app" in EthercatMCApp.