EthercatMCAxis.cpp/readBackHoming(): Use setIntegerParam
There is a difference between
  pC_->setIntegerParam(axisNo_, pC_->EthercatMCHomProc_, homProc);
and
  setIntegerParam(pC_->EthercatMCHomProc_, homProc);

The first one calls the base class, and the second one the overloaded one
from EthercatMCAxis.cpp

For the homing procedure we need the overloaded one, to set
motorNotHomedProblem_ to the right value.

Use it and fix the problem that "Axis not homed" did not work, whenever
the homing procedure was only read back from the controller.