Correct setSim in config file reader
When we run a config file with the "setSim" keyword against a real
controller (not the simulator), the reading of the config file failed.

Set the status to asynSuccess when not running against a simulator.

While at it, correct a wrong printout (which said thatgetFeatures()
failed instead of readConfigFile().

And correct a wrong indentation.