Introduce new error code "Not found"
When an axis is configured. but nMotionAxisID can not be found, the poller tries to
find it forever. Each lookup will loop over 10 possible ADS ports, blocking the
ASYN port for 10 seconds.
This is not ideal.
In a first step, use nMotionAxisID == -2 as "axiis not found", and don't try to
find it any more.
The next step could be to reduce the number off possible ADS ports,
that will be done in the next commit-