simulator: Allow to set same scaling even if power is on
When the IOC is restarted, the scaling may be pushed into the simulator.
Changing the scaling is not allowed in the real system,
when the amplifire is on.

Do an exception for the simulator:
allow to set the scaling to the same value as before, even when the
amplifier is on.

In the end, this could be better fixed in the driver.