Optimize the logging: modulName only when needed
The asynPrintf is able to print file:line, if enabled.
Once file:line is enabled, we don't need the modulName to be printed.
When the axis is created, look at the traceInfoMask.
When file:line is printed by asyn, set modulName to "".
At the same time, rename moduleName to modNamEMC to have it global
and used by both EthercatMCAxis and EthercatMCHelper.

Change even startup/EthercatMCController.cmd and don't print the thread.