simulator: Keep the position when parking position os set
When the IOC is started, the parking position may be set using a config file
with a line like this:
setSim fMotorParkingPosition=0.1

Today this will set the parking postion and the actual position.
As a result, the position is always back to parking, when the
IOC running against the simulator configureing the parking position restarts.

This is not ideal, we want to keep the position unchanged.

Solution:
Only if a new, different parking position is set, change the actual position
of the simulated motor.