Add mov2(): All commands to MCU in one go
Optimize to time needed to perform a move():
So far there 7 transactions used to set up all parameters,
each of them needed a round-trip over the TCP/IP connection.
Optimize by sending all commands in one go.
This is done in mov2().
Note: Because the ADSPORT is needed to be prefix for each and every
command, a longer buffer is needed.

MAX_CONTROLLER_STRING_SIZE 256 was increased to 1024 in synMotorController.h