Simulator: Add usleep() and TC (WIP)
In order to simulate a "highly-loaded" TCP/IP connection,
add a usleep() in the simulator.
Before processing the command, a sleep is executed (if configured).
Having a sleep in motor 2 will affect the polling, and even effect
motor 1 (and all other motors).
Add a test case for this, there should be more code later on.