EthercatMCAxis: simplify move()
Simplifiy the code in move():
The acceleration is given from the generic driver in steps/sec/sec
and not in "seconds to velocity".
Note that the "step size" is MRES in the motorRecord.
Before the values are send to the controller, they are multiplied
with an internalstepSize.
In this way a database can be contructed in um (MRES=0.000001) and
the controller configured to use mm (drvLocal.stepSize=0.001).

This can be useful for some people and confusing for others at the
same time.