Simulator: add setPositionJitter (flickkering encoder)
Make it possible to simulate a flickering encoder:
Each poll returns a different measured position.
The (measured) position is toggling around a value with a +/- delta.
The delta is configureable.
The IOC will get a new measured position with each poll(),
and the motorRecord will be processed after each poll().