A moving axis while powere off is not commanded
When a motor is powered off, don't report is as "moving",
when the position is changed.

A "moving" motor will synchronize the "RBV" into "VAL" at the
end of the move, this is treated as an "external" move in the
motorRecord.
Since there is no real movement involved, don't set the moving
bit in the driver.