EthercatMCAxis: Print fPosition, nCommand and nCmdData at first poll
In order to improve the connection of an IOC to the controller,
print out the values of fPosition (the commanded position) together
with the values of nCommand and nCmdData.
In case that the last move had been a commanded move, fPosition
can be used instead of fActPosition. (But that is for a new commit)