Improve the errorMessage handling
The varianle sErrorMessage may have a useful content, worth to be
shown to the user.
On the other hand, we may have "well-known" ErrorId, which are
developped outside this project, with a stable numbering and meaning.
When we have a well known meaning, show the "hard coded" text.
Else show the text from the controller, expecially for ECMC.