readbackEGU
When the motor module supports the "readbackEGU", use it.
Both motor- and encoder positions are send in EGU to the
driver/motorRecord.
Add a new parameter, scaleFactor.
It allows to have the controller in mm and e.g. the record in um.
Not that we want to use it that way, but as the code for scaling
is already there, simply keep it.
However, the stepSize option is ignored when "readbackEGU" is used.
This should allow to use old start scripts together with old versions
of the motor module.