TC110: Detect when "-HomProc" is not available
Depending on the loaded datbases (template file) the PV to set
the "homing procedure" may not be available.
Detect when it is missing, and stop try to run further tests
(which will only wait for timeouts, without doing anything useful).

As the first test, try to read "$(P)$(M)-HomProc", and if that
fails, assert in the first test case, and disable the others.