EthercatMCaxisExpert.opi: Add ACCS, VELO, ACCL
Add fields to the screen, especially VELO and ACCL.
The Dec-RB field is not show any more, (to give some space),
we use Acc == Dec anyway.

The ACCS field is a new field in the motorRecord, show it if the
motorRecord has a version which is known to support it.