Retry when TwinCAT reports 'State timout'
Sometime the epics-com-module in TwinCAT runs into a timeout,
and the answer is something like
"Error 1 : State timout. To long time in one state."

This should be fixed in the TwinCAT code, but until this happens,
do a retry.

Add some more printouts to be able to debug the problem.