EthercatMCAxis.cpp; Improve the printing of config readback
Many values/parameters are configured inside the motion controller.
Already today they are read into EPICS PVs, but the values are
not in the IOC log file.
Add a print in the log file, so that the log file can be send via
email for futyher analysis.

Introduce updateCfgValue() to do the printing.