test/motor_lib.py: Make setSoftLimitsOff() more robust
The is a possible race condition when the read-only softlimits
are switched off, followed by switching off the motorRecord soft limits.

First the controller needs to be instructed to disable the soft limits
(Not that this is only possible in debug mode), then the new information
must reach the motorRecord, and once that is done, the motorRecord
soft limits mut be disabled by setting HLM = LLM = 0.0.

This sometimes fails, and LLM stays != 0.0
As a result, we can not JOGR against the low limit switch and activated it.

Make a retry loop in motor_lib, to make sure that the limits are off and
let e.g. TC 132 pass.