Add EthercatMCmcUTCtime.template and UTCtime.proto
Add a template and a protocol file to retrieve the UTC from the
motion controller unit.
A special SW is needed in the motion controller to provide the variable
inside the project.
And we need a 1 Hz puls from e.g. a timing system connected to input 1
of the EL1252-0050 to run the test.