EthercatMCIndexerAxis: Allow axis to be created with options
To be able to test the statusWord of the PILS protocol,
allow an IndexerAxis to be created with options.using the iosh command.

Becuase the indexer does not create those axes, move the writing
of the name of the reason bits 27..24 into the axis code.

Temporary allow the indexerAxis to be used with Gvl,
that code will disappear later.