Boy/motorxx.opi: Show a Rst-button when -Err is set
Make it possible to do an "ErrRst" from the user screen
whenever the -Err bit is set, indicating an error from the controller.

Move the "More" and the "STOP" button more into the corners.
Re-order the source code so that those 2 elements are last,
as they have the highest y-value.

Place the Rst button in between them and hide it until
the -Err bit of the axis is on.