EthercatMCAxis.cpp: Add axisFlags, powerAutoOnOff, powerOff/OnDelay
Make it possible to fine-configure the autopower-settings more fine-grained
when the axis is created.

In a more typical EPICS application this would be done with records,
and that is possible even here.

However, for the moment there is no need to records for all the auto-power
settings (they are in the "Debug" db file).

While there, add a config option "axisFlags" which overrides the flags.

All in all we have 4 new options:
axisFlags, powerAutoOnOff, powerOffDelay, powerOnDelay