run-EthercatMC-ioc.sh: Improve handling of local AMS NetId
When running against a system via ADS, we need to make sure that the
local AMS NetId is correct, otherwíse there will be no answer send
to the IOC.
Demand that the  local AMS NetId is provided on command line.

If it is not given, loop through the interfaces with help of
`ip addr` or `ifconfig` and suggest the local IP addresses to be used,
appended with ".1.1".

While there, convert TAB to space.