EthercatMCaxisExpert.opi: Add a (hidden) StatusCode field
EthercatMCaxisIndexer.opi should be removed in the long run.

Start to integrate displayed fields from EthercatMCaxisIndexer.opi
into EthercatMCaxisExpert.opi.

The first step is to show the StatusCode in the expert screen,
when it is initialized - otherwise it is hidden.

As a result, the StatusCode PV is moved from
EthercatMCindexer.template into EthercatMC.template.