test/simulator: TC141 broken for motor 4
The simulator put an "axis number" (axis_no) into the logfile,
when moveVelocity was commanded and started.

To make the logfile match, the test script used a hard-coded
value, "axis_no=1" to match that log line.

This does not make much sense, the log file belongs already to
the motor (the motorPV to be exact) and the test script has no
idea about internal axis numbers in the controller.

So remove the axis_no and make TC141 pass for motor 4, IOC:m4.