EthercatMCIndexerAxis: Don't snoop an functions as parameters
The parameter interface defines paramters and functions.
If a parameter is changed from outside, we snoop and put that
changed parameter into the paramlist at the IOC.

This does not work well for function 142:
The parameter is read back, put into the "-CfgJVEL" PV,
propagates into motorRecord.JVEL and triggers a second
"moveVelocity" command.

TC 141 fails.

Solution:
Don't snoop on functions.