Simulator; Add moveRelative command (simulator only)
The PILS spec does not specify a "move relative" command,
since that can typically achieved with an "move absolute" command.

Use the current position, add the delta, and move to that position.

However, test case 141 is testing that the motorRecord commands
relative movements (under certain circumstances).

Until today the relative-absolute conversion had been done
in the driver on the IOC (and still is),
In oder to be able to change this, add a command like
"Sim.M1.moveRelative=8.0".

To make it possible to specify acceleration and velocity ussing
the parameter interface, save those in hw_motor.c using
setNxtMoveAcceleration() and setNxtMoveVelocity().