Indexer: parameterFloatReadBack() needs an initial parameter
When the IOC is connected to the MCU, different "config parameters"
are read and put into different PV in EPICS.
This is done using parameterFloatReadBack().
The same function is used, to update the "read back" PVs.

Don't mix those up!

Example:
"EthercatMCCfgVELO_" should only be read once, when the connections
is established.

But "EthercatMCVel_RB_" should be updated regulary, as it is a runtime
parameter.

This was detected by TC 141.