test: Speed up 141
Decrease the sleep time from 3 to 1 seconds.

While there, add a call to
lib.initializeMotorRecordSimulatorAxis() which was previously done
test case in the 13x range.
(The 13x had been renamed, so that it is hard to follow what exactly
 had been done here. `git log --follow` may help)

And, as a bonus, add printouts of line numbers, just in case