Indexer: Add AddPollNowParam()
In order to improve the readback of paramters in the MCU via
the indexer, add the method AddPollNowParam().
Calling this will add this parameter index to the list of polled
parameters, which is maintained for each axis.

Each axis may expose it's own set of parameters.

This method is not used yet, that will be part of the next commit.