EthercatMCHelper.cpp: Avoid over-long printing into msgTxtFromDriver
gcc warned about an overlong string being writtin into errBuf,
which will be displayed on the UI.

Reduce the number och characters printed here, they will no fit on
the screen anyway.
Move printing of the additional information e.g. inString_ into
asynPrintf, so that we can se it in the IOC log.