ethercatmcIndexerAxis Bug fix when changing "read only soft limits"
When the read only soft limits are changed, the method udateMotorLimitsRO()
must be called to inform the motorRecord abiut the new "read onlz soft limits".

However, this function needs to have the actual values in the parameter library,
and the old code did not update the parameter library.
Fix this (detected by test case 900)