Indexer: Handle errorID for PILS 5008 motors
The PILS device 5008 for a motor does not include the errorID.
(The 5010 does). In order to get the errorID into the EPICS IOC,
we use a "Simple discrete input (32 bits)" (type code 1202) with the
name "errorID" which must be located after the 5008 device.

Add function setErrorIdOffset() to tell the axis where the device
can be found in the process image, and us it in poll().