ethercatmcIndexerAxis: Fix wait for power on
The setClosedLoop() function was buggy:
The value send to the MCU is 1.0 for “closedLoop=0, power off”,
And 0.0 for “closedLoop=1, powerOn”.

The parameter library was updated with value.
This gave problems when the amplifier should be turned on,
the parameter library returned “1”, but the hardware was still “0”.

Commanding a move now failed, and the axis goes into error state.
While there, add some printouts when “auto power” is changed.