test: JOG_LS: Move really away from LS
The axis 1 in indexer has a RDBD of 2.0, and moving away
from limit switch dows not work as expected.
The move from e.g. 176 (the position of the limit switch)
to 175 (the high soft limit HLS) does nothing.
The command gets stuck in the motorRecord because we are inside RDBD