Add time stamping of PV with DC clock
Read out the DC clock, when commint from the MCU via PILS, and
timestamp a new PV with that timestamp.

Add a ne PV "-PosAct" that hooks onto the existing asyn parameter
"MOTOR_ENCODER_POSTIONE" and set the TSE field to -2.

In the poll() method of the indexer, retrieve the DCtime, and set it onto
the asynport, so that all records exept the motor record will have this
time stamp.

(Side note: The format of DC time will change in the next version of
 MCU firmware, so the handling of DC time needs to change.
 But the principle will stay the same)