More DCclock time stamping
Implement the DCclock-delivery:
We get the 64 bit counter from the MCU "as is".
The counter counts nano seconds, 0 is 2000-01-01 00:00
The camonitor gives:

LabS-MCAG:MC-MCU-07:m1-PosAct  2020-04-01 16:40:47.450054 0.81
LabS-MCAG:MC-MCU-07:m1.RBV     <undefined> 0.81

And after a movement:
LabS-MCAG:MC-MCU-07:m1.RBV     2020-04-01 16:39:13.538580 1.2
LabS-MCAG:MC-MCU-07:m1-PosAct  2020-04-01 16:40:57.030057 1.2

We need to investige, why this is as it is.