Simplify the EOS handling
The ASCII based commicunication needs input- and output-eos.
For traditional reasons, those had been configured in st.cmd
And now the binary (ADS) communication does not need the eos
at all, so that the code sets the eos to "" for each and every transaction.

It's time for a cleanup:
Remove the setting of eos from the start script, and do it once and forever
(if needed) in the controller code.