ethercatmcADS.cpp: Improve printing in writeReadBinaryOnErrorDisconnect()
The handling of print-levels was not ideal in that function:
When the MCU did not responde, printing was suppressed after MAXCNTADSSTATUS,
regardless of the tracelevel.
So even if
asynSetTraceMask("$(ASYN_PORT)", -1, 0xFF)
was used in the startup script, nothing was printed.

Modify the logic, so that printing is always called.
Change the code so that the tracelevel is raised, if needed.