ethercatmcAxis::setPosition(): Don't return asyError for simulator
When MRES is changed, the motorRecord issues a LOAD_POS, resulting in
a call to setPosition().
In generally, this is forbidden and returns asynError,
unless the homing procedure is "manual set position".
Loosen that error handling somewhat:
Some test cases run against the simulator, so don't return asynError here.
Silently ignore the position setting.