test/TC900: Correct calculations for MRES != 0
The motorRecord soft limits (DHLM, DLLM) are limited by the
"controller read only soft limits".

900_Simulator-Softlimits.py changes both MRES and the controller soft limits
and wait until the new values have rippled throug the controller, the driver
and all records in a "polling loop".

That loop took 5 seconds because the calculation was wrong.
The test case worked.

Improve the situation:
- correct the calculation of DHLM/DLLM (need MRES)
- Raise an exception when the busy loop is still not happy
- Rename InitLimitsWithC() into InitLimitsWithROlimits,
  InitLimitsNoC() into InitLimitsNoROlimits()
- Minor updates in the logging statements