test/900: Harden the setup of "read only softlimts"
Sometimes the enabling of a soft limt does not go through.
The python scripts writes 1, but the MCU gets a "0".
This could be a race condition:
The low soft limit is enabled, then all enabled bits are read back from the
controller, and the the high soft limit should be enabled.
For the moment, I suspect that the "readback from controller" overwrites
the new setpoint in the parameter-library.

In principle, it would be better to use different records for
readback and setpoint.
As disabling soft limits is only done under testing, and not
for a normal user, we can live with a workaround in the test script.