De-couple CfgDHLMXXX and CfgDHLMXXX_RB
Use 2 different parameters for the "readback" of the MCU softlimits
and the values that are "pushed" into the MCU.
Under normal user-operation no soft limit is updated by the IOC,
but for testing we sometimes disable the soft limits or change the values.

Make clear, which is which and use separate parameters.