Indexer: Improvements around soft limit handling
Improments when soft limits are set in the indexerAxis:
The parameter change like this
setIntegerParam(1 ethercatmcCfgDHLM_En)=0

caused "ethercatmcCfgDHLM_En_RB_" to be set immediatly.
This should no longer be used, we now wait for the poller to pick
up parameters from the controller and the we see a line like this:

updateCfgValue(1) dhlm_en=1

This "slow and save" polling of the readback variables may take some
time and slow-down the testing, but that is OK
(and should already be covered by the test scripts).
 message for your changes. Lines starting

Other improvements:
- poll(): Don't print the targetPostion (it is 0 for the moment)
- paramInterface: Print the status in text (like "PARAM_IF_CMD_DONE")