test/910: Proper axisMr.waitForMipZero()
The timeout for waitForMipZero() was too short, and sometimes
the test failed. (This was hard to find, since the old backlash
was still ongoing, when the new test was started).
Anyway, now that we have the exception-raise in AxisMr.py,
it is clear that the timeout was too short.
Fix it.
While there, remove the extra wait() calls, which are not needed.