IndexerAxis: Don't bail out if velocity/acc cant be configured
When the contoller does not expose velocity and/or acceleration,
the indexerAxis did try to write them anyway, which failed.
And as a result, the whole move() failed.

Loosen the condition for failing:
When reading of ethercatmcVel_RB_ returns "asynParamUndefined",
this parameter had never been read for this axis, because it is not
"exposed". In this case, don't try to write it either.
Do the same for ethercatmcAcc_RB_