Indexer: Handle "simple output device"
When the "homing procedure" should be changed in the PLC via PILS,
an "output device" is needed.
Add handling for a "Simple discrete output (32 bits)".
Despite it's name, it has a "Current value" and a "Target value".
The "Current value" could be used as HomProc-RB, but this is
not done yet.