.travis.yml: New set: ess.set
Run the complete test suite with the new "ess.set".
ess.set reflects what we have deployed so far -
and that is not always the latest version. E.g. asyn is at R4-33.

Correct the order of the modules:
The latest asyn needs calc be build before asyn.

Other changes:
  ci-scripts: Use v3.0.2
  .travis.yml: Enable verbose mode, VV=1
  .travis.yml: traviscache2/ -> traviscache/
    Change the directortoy and the cache directory