ethercatmcAxis.cpp: Limit switches and mov2() for classic motorRecord
When a limit switch is activated, the classic motorRecord will assume that
the motor has fully stopped.
Our motors need a ramp-down, and report mvnNRdyNex = 0 later.
Workaround:
"hide" the active limit switch to the record, as long as the motor is still
"busy" and mvnNRdyNex is 1.

While there, #ifdef out mov2() when compiling against motor which has the
"short" MAX_CONTROLLER_STRING_SIZE