test/indexer: update setNxtMoveVelocity in MOVE_VELOCITY
In order to get the upcoming test case 950 through,
write the commanded velocity into
setNxtMoveVelocity(motor_axis_no, fValue);
when PARAM_IDX_FUN_MOVE_VELOCITY is used.

Not sure, if this is the right thing to dom though.
Time will show, if we need to remove this and find a better way
to test the JVEL in TC 950.