Add PV Freeze-Offset-Visible, -FoffVis
If "setPosition()" is allowed is today determined by the homing procedure.
Decouple this relation and introduce a new PV, ($M)-FoffVis
This allows to "manually calibrate" an axis via ".FOFF" = 1,
".SET" = 1, which results in a call to setPosition()
as well as running a homing sequence.

Note:
Today the indexer already is well-prepared:
There are 2 independent functions:
 133 "Home in default direction"
 137 "Set position":
      moves zero-point so that current value becomes the given argument

Note2:
More work to be done with the simulator
(And may be the "classic" ethercatmcAxis, if someone needs it.