Simulator: setPosition() uses MotorPosSetPosOffset
Major rework of setPosition():
The position will be pushed into the controller/axis.
Keep track of the HW: When running into limit switches,
the simulated HW must use a "HW position" to determine if a limit
switch was hit, and not the MotorPosition communicated to the IOC
(and used everywhere).

The same fot the simulated encoder: If the position is changed,
the encoder stays at it's value.
The MotorPosSetPosOffset can be reseted by "homing" the motor.