test: 900...py: fix InitLimitsWithROlimits()
When testing different combinations of controller-read-only-soft-limits
the new settings may need some time to ripple from the controller through
the poller into the ".DHLM" and ".DLLM" fields of the motorRecord.

Fix the "polling" task: Set up the values in the controller once
(and only once).
Then "loop and poll" ".DHLM" and ".DLLM" fields.