ethercatmcIndexer.cpp: Initial update of ethercatmcCfgDxLM_En_
There are 2 types of PV about the enabled soft limits in the MCU:
ethercatmcCfgDHLM_En_RB_ (What is in the controller and is reported back)
ethercatmcCfgDHLM_En_    (The IOC can switch off the limits. For testing only,
                          but not for production).

When the IOC was started, the ethercatmcCfgDHLM_En_ was not initialized,
leaving the value at 0, even if limits had been active insinde the controller.
Do an initial only readback to get them in sync.

Note: The same applies for the "Low" PV as well