test/indexer.c: Correct handling of AUX bits
The simulator code had some confusion between the device number and
the motor number when calculating the status of the 5080/5010 devices.

The "homed" status belongs to the motor number, but the AUX bit names
belongs to the device number.
(In the C-code: "devNum" vs "motor_axis_no").
Fix this mix up, and pass both devNum and motor_axis_no into
indexerMotorStatusRead5008() and indexerMotorStatusRead5010().

While there, pass if the number of AUX bits, and fix an off-by-one error,
which didn't handle bit 7 at all in the 5080 devices.