Indexer: Implement output device 1604
An output device should be handled in setIntegerParam().
Add the code here.
Refactor pilsAsynDevInfo_type so that it is more clear where the
payload is.
Note that an output device in the MCU has allways an "actual" value
associated with it, and now only the output part is used.
One the long run there should be 2 asynParameters created:
one for the "readback" and one for the "target" value.
That needs to go into different commits.