Indexer: Refactor into newPilsAsynDevice()
Refactor addPilsAsynDevInfo() into
newPilsAsynDevice() and
addPilsAsynDevList().

Call newPilsAsynDevice() when a PILS device is found in the indexer,
which figures out the type (input/output, integer/floating point).

Create an asyn parameter, and once that is done, "register" the
asyn parameter number ("function") in the poll list for input-
and ethercatmcIndexerAxis::setIntegerParam() for output-devices.
No support for 64 bit integer output devices yet, and no suppoert
for ethercatmcIndexerAxis::setDoubleParam() either.