Introduce indexerDisconnected()
When the comminication is broken, call a new method indexerDisconnected()
The purpose is to do all needed cleanup.
Additionally we want to set an alarm status on the additional
asyn parameters and their records, that will be done in the next commit.

While there, rename initialPollIndexer () into indexerInitialPoll()
and pollIndexer() into indexerPoll() to have a more consistent naming.