Add setAlarmStatusSeverityWrapper()
Make it possible to set alarm status and severity by using a wrapper
method.
Hide the details of all EPICS base (epicsAlarmCondition epicsAlarmSeverity).
Hide the details of the asyn wrapper wrapper as well and use
setAlarmStatusSeverityWrapper(axisNo,
                              function,
                              [asynSuccess|asynDisconnected])

The new method is not used yet, this will be done in the next commit.