ethercatmcdebug: Intialize -CfgDxLM, -CfgDxLM-En
The ethercatmcdebug.template exposes PV to switch off or change
the value for the softlimits inside the controller.
This is for debugging only, not for production.
It allows to test the limit switches.

Since these records are output records alltogether, initialize them
with the value(s) from the controller, once the IOC connects to the
controller.