Add DbgStrToNC, part 1
Prepare a Streamdevice-ish possiblity to talk to the NC block in the MCU
directly.
Streamdevice uses the octet interface to talk to an asyn poert driver.
Add code to handle writing (but don't use it via Streamdevice yet)
Instead, add a stringout record, which needs to be feed with the
ADS command.
Note: This is for testing only, mainly to test the parameter handling
inside TwinnCAT, this is not for production.

TODO:
- Implement the writing of double values.
- Implement test code.
- Implement a range check: The axis number must be checked if it matches
  the axisID.