ethercatmcAxis: Improve homProc handling on re-connect
Need to improve the handling of especially ethercatmcHomProc_RB_,
when the connection to the controller is lost and re-established.

If the homing buttons are visible is determined by ethercatmcHomeVis_
and the FOFF but needs ethercatmcFoffVis_.
Once the connection is lost, HOMF, HOMR and FOFF buttons are hidden in CSS:
The status of all 3 had been set to asynDisconnected.

Make sure that the status is set to asynSuccess after the re-connect and
the homing procedure is re-read from the controller.
Otherwise they will never re-appear.

Improve the handling, when the homing procedure was given as a parameter.
They will disappear and re-appear, but no attempt to read them from
the controller should be made.