ethercatmcIndexer: Fix indexerParamRead() again
Remove some confusion/bugs when using the param interface:
The READ_CMD was issued in every loop - both confusing
and wrong. At the end the read succeded after doing some extra rounds.

Fix the loop to handle bot collisions (with other hosts) and handle
the processing time in the MCU.
TODO: Do the same for write.