simulator: Add get/setMaxAcceleration()
The PILS protocol demads to limit the acceleration to a max value -
strictly speaking it demands that the accleration is at a useful value.

Implement a range check to catch too high values.
Note:
More checking (for negative values =) may be needed later, let's start
to check if the EPICS code handles this nicely.